+++ src/propylean/instruments/measurement.py
@@ -1,4 +1,4 @@
-from propylean.generic_equipment_classes import _EquipmentOneInletOutlet
+from propylean.equipments.generic_equipment_classes import _EquipmentOneInletOutlet
 
 
 class FlowMeter(_EquipmentOneInletOutlet):
